refactor: properties 파일 global 디렉토리로 이동
