+++ src/main/java/com/aliens/backend/global/property/S3UploadProperties.java
@@ -1,4 +1,4 @@
-package com.aliens.backend.uploader;
+package com.aliens.backend.global.property;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
