+++ src/test/java/com/aliens/backend/uploader/AwsS3MockConfig.java
@@ -1,5 +1,6 @@
 package com.aliens.backend.uploader;
 
+import com.aliens.backend.global.property.S3UploadProperties;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.AnonymousAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder;
